reading course info into objects
The xlrd just isn’t working that well with the course info so I turned
it into a txt file then read the data into objects.
***Each Course stores a time = [start, end]….The times are in 24 hour
clock/ military time to make for easier comparison. Also the minutes
are rounded to a decimal to make for easier comparison as well. I set
it up so that you can take the start and end times of a class(that a
professor has) and see if either of those is >= start of a diff. class
and <= end of a diff. class….(then that means the teacher can’t teach
the diff. class)